add updating state
